requirements.txt
asgiref==3.8.1
Django==5.1.1
django-taggit==6.0.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
python-decouple==3.8
sqlparse==0.5.1
